ported the Exportable module from YourSAy for better csv generation
This solves the problem of csv generation on large tables
potentially crashing servers due to having to render hundreds
of thousands of partials. It also provides more consistent
formatting of things like dates, booleans, and has_many associations,
and is extensible, similarly to the old method, but with `csv_format_<type`
class methods on the model rather than with partials.