Remove in-memory nav config cache
workaround for simple navigation bug. rendering nav causes nav options to be deleted, including :if lambda procs. AIT uses nav :if lambda procs to hide nav items depending on whether a user is logged in. AIT renders two navigations (desktop and mobile). The second nav to be rendered (mobile) does not include the :if lambda procs. So the mobile nav shows items it shouldn't.

This change removes the in-memory cache of navigation configuration, so options are regenerated each time navigation is rendered.