Remove NavItem eval attributes
NavItem previously supported runtime eval for customising navigation
menus, but this was inherently un-cacheable and has not been possible
to edit these fields since v2.2.0 when caching was added (2016).

This commit adds a migration to remove the attributes and the runtime
eval implementation, which allows for a significant code cleanup.