Patch ActionView to de-duplicate partial paths inside admin
When an object is rendered inside an admin controller, we want to look
for the admin partial for rendering that object.

We've filed a Rails issue about this use case, but it's unlikely to
be merged any time soon. Adding this patch makes it possible to
define polymorphic partials for nested admin controllers without
repeating the module directories.