HasAttachments controller support for saving attachments
When an admin controller accepts files for upload, it's usually
necessary to save attachments before rendering errors so that the
same attachment doesn't need to be uploaded multiple times.

This could also be solved by adding the direct uploads JS, but we
have not yet decided whether to include that by default in new Koi
applications.